[CET-6660] Move select to standalone file
