+++ src/components/ReportsPage/HeatmapPage/HeatmapFiltersModal.tsx
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 import React, { useMemo, useState } from 'react';
-import { Box, Button, Dialog, DialogActions, DialogContent, DialogTitle, FormControl, InputLabel, MenuItem, Select } from '@material-ui/core';
+import { Box, Button, Dialog, DialogActions, DialogContent, DialogTitle, MenuItem } from '@material-ui/core';
 import { StringIndexable } from './HeatmapUtils';
 import { HomepageEntity } from '../../../api/userInsightTypes';
 import { sortBy, uniq } from 'lodash';
 import { Clear } from '@material-ui/icons';
+import { ModalSelect } from './HeatmapFiltersSelect';
 
 export interface ScoreFilters {
   serviceIds: number[];
@@ -38,41 +39,6 @@ interface HeatmapFiltersModalProps {
   filters: ScoreFilters;
   setFilters: (scoreFilters: ScoreFilters) => void;
   entitiesByTag: StringIndexable<HomepageEntity>;
-}
-
-interface ModalSelectProps<T> {
-  name: string,
-  value: T[],
-  onChange: (value: T[]) => void,
-  onReset: () => void,
-  options: React.ReactNode
-}
-
-const ModalSelect = <T,>({ name, value, onChange, onReset, options }: ModalSelectProps<T>) => {
-  return (
-    <Box display="flex" flexDirection="row">
-      <FormControl variant="standard" fullWidth>
-        <InputLabel>{name}</InputLabel>
-        <Select
-          multiple
-          value={value}
-          onChange={event => onChange(event.target.value as T[])}
-        >
-          {options}
-        </Select>
-      </FormControl>
-      {value.length > 0 && (
-        <Button
-          onClick={onReset}
-          variant="text"
-          aria-label={`Clear ${name.toLocaleLowerCase()} filter`}
-          title={`Clear ${name.toLocaleLowerCase()} filter`}
-        >
-          <Clear />
-        </Button>
-      )}
-    </Box>
-  )
 }
 
 export const HeatmapFiltersModal: React.FC<HeatmapFiltersModalProps> = ({ filters, setFilters, entitiesByTag }) => {
